Update `RevitTest` version to `1.7.1`

+++ ricaun.DA4R.NUnit.Console/DA4RTestService.cs
@@ -56,7 +56,7 @@ namespace ricaun.DA4R.NUnit.Console
             string forceLanguageToRevit = null,
             bool forceToOpenNewRevit = false,
             bool forceToCloseRevit = false,
-            int timeoutMinutes = 0,
+            double timeoutMinutes = 0,
             params string[] testFilters)
         {
             LogApplicationInfo();
